requirements.txt
torch
numpy
pandas
pillow
webcolors
django
djangorestframework
markdown
fastapi
uvicorn
